+++ lib/src/form/field_bindings/ArrayOfReferencesFieldBinding.tsx
@@ -18,7 +18,7 @@ import { getReferenceFrom } from "../../core/util/entities";
 type ArrayOfReferencesFieldProps = FieldProps<EntityReference[]>;
 
 /**
- * This field allows to select multiple references.
+ * This field allows selecting multiple references.
  *
  * This is one of the internal components that get mapped natively inside forms
  * and tables to the specified properties.